% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeclare.R
\name{as_pdeclare}
\alias{as_pdeclare}
\alias{as_pdeclare.tbl_df}
\alias{as_pdeclare.grouped_df}
\alias{as_pdeclare.data.frame}
\alias{as_pdeclare.list}
\title{Coerce to a pdeclare tibble}
\usage{
as_pdeclare(x, .i = NA, .t = NA, .d = 1, .uniqcheck = FALSE, ...)

\method{as_pdeclare}{tbl_df}(x, .i = NA, .t = NA, .d = 1,
  .uniqcheck = FALSE, ...)

\method{as_pdeclare}{grouped_df}(x, .i = NA, .t = NA, .d = 1,
  .uniqcheck = FALSE, ...)

\method{as_pdeclare}{data.frame}(x, .i = NA, .t = NA, .d = 1,
  .uniqcheck = FALSE, ...)

\method{as_pdeclare}{list}(x, .i = NA, .t = NA, .d = 1,
  .uniqcheck = FALSE, ...)
}
\arguments{
\item{x}{A data frame, tibble or list}

\item{.i}{Character or character vector with the variable names that identify the individual cases. If this is omitted, \code{pdeclare} will assume the data set is a single time series.}

\item{.t}{Character variable with the single variable name indicating the time. \code{pmdplyr} accepts two kinds of time variables: numeric variables where a fixed distance \code{.d} will take you from one observation to the next, or, if \code{.d=0}, any standard variable type with an order. Consider using the \code{time_variable()} function to create the necessary variable if your data uses a \code{Date} variable for time.}

\item{.d}{Number indicating the gap in \code{t} between one period and the next. For example, if \code{.t} indicates a single day but data is collected once a week, you might set \code{.d=7}. To ignore gap length and assume that "one period ago" is always the most recent prior observation in the data, set \code{.d=0}. By default, \code{.d=1}.}

\item{.uniqcheck}{Logical parameter. Set to TRUE to perform a check of whether \code{.i} and \code{.t} uniquely identify observations, and present a message if not. By default this is set to FALSE and the warning message occurs only once per session.}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
WARNING: FOR NOW, BE AWARE THAT AFTER USING AS_PDECLARE, MOST FUNCTIONS WILL NOT PRESERVE THE ATTRIBUTES.
}
\details{
This function coerces a tibble, data.frame, or list to a pdeclare tibble by adding three attributes to it:

\itemize{
  \item \code{.i}, a character or character vector indicating the variables that constitute the individual-level panel identifier
  \item \code{.t}, a character vector indicating the time variable
  \item \code{.d}, a number indicating the gap
}

Note that pdeclare does not require that \code{.i} and \code{.t} uniquely identify the observations in your data, but it will give a warning message (a maximum of once per session, unless \code{.uniqcheck=TRUE}) if they do not.
}
\examples{
data(SPrail)
# I set .d=0 here to indicate that I don't care how large the gap
# between one period and the next is.
# If I want to use 'insert_date' for .t with a fixed gap between periods,
# I need to transform it into an integer first; see time_variable()
SP <- as_pdeclare(SPrail,
  .i = c("origin", "destination"),
  .t = "insert_date",
  .d = 0
)
is_pdeclare(SP)
attr(SP, ".i")
attr(SP, ".t")
attr(SP, ".d")

data(Scorecard)
# Here, year is an integer, so I can use it with .d = 1 to
# indicate that one period is a change of one unit in year
# Conveniently, .d = 1 is the default
Scorecard <- as_pdeclare(Scorecard, .i = "unitid", .t = "year")
is_pdeclare(Scorecard)
}
