% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeclare.R
\name{build_pdeclare}
\alias{build_pdeclare}
\title{Low-level constructor for a pdeclare object}
\usage{
build_pdeclare(tbl, .i = NULL, .d = 1, .t = NULL,
  .uniqcheck = FALSE)
}
\description{
`build_pdeclare()` creates a `tbl_pd` object with more controls. It is useful
for creating a `tbl_pd` internally inside a function.
}
