% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdeclare.R
\name{is_pdeclare}
\alias{is_pdeclare}
\title{Function to check whether an object has been declared as panel data}
\usage{
is_pdeclare(.df, .silent = FALSE)
}
\arguments{
\item{.df}{Data frame or tibble}

\item{.silent}{Set to TRUE to suppress output reporting what the panel identifiers are. Defaults to FALSE}
}
\description{
Checks whether a data set (\code{data.frame} or \code{tibble}) has been assigned panel identifiers in the \code{pmdplyr} format. If so, returns those identifiers.
}
\examples{

data(Scorecard)
Scorecard <- as_pdeclare(Scorecard, .i = "unitid", .t = "year")
is_pdeclare(Scorecard)
}
