% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_consistency.R
\name{mode_order}
\alias{mode_order}
\title{Calculate the mode, and use original order to break ties}
\usage{
mode_order(x)
}
\arguments{
\item{x}{Vector to calculate the mode of.}
}
\description{
\code{mode_order()} calculates the mode of a vector, mostly used as the default \code{.resolve} option in \code{fixed_force()}
}
\details{
In the case of ties, the first-ordered value in the vector wins.
}
\examples{

x <- c(1, 2, 2, NA, 5, 3, 4)
mode_order(x)

# Ties are broken by order
x <- c(2, 2, 1, 1)
mode_order(x)
}
