% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_within.R
\name{panel_calculations}
\alias{panel_calculations}
\alias{within_i}
\alias{between_i}
\title{Functions to perform standard panel-data calculations}
\usage{
within_i(.var, .df = get(".", envir = parent.frame()),
  .fcn = function(x) mean(x, na.rm = TRUE), .i = NA, .t = NA,
  .d = NA, .uniqcheck = FALSE)

between_i(.var, .df = get(".", envir = parent.frame()),
  .fcn = function(x) mean(x, na.rm = TRUE), .i = NA, .t = NA,
  .d = NA, .uniqcheck = FALSE)
}
\arguments{
\item{.var}{Vector to be transformed}

\item{.df}{Data frame or tibble (usually the data frame or tibble that contains \code{.var}) which contains the panel structure variables either listed in \code{.i} and \code{.t}, or earlier declared with \code{pdeclare()}. If \code{tlag} is called inside of a \code{dplyr} verb, this can be omitted and the data will be picked up automatically.}

\item{.fcn}{The function to be passed to \code{dplyr::summarize()}. \code{x - .fcn(x)} within \code{.i} is the within tranformation. \code{.fcn(x)} within \code{.i} minus \code{.fcn} overall is the between transformation. This will almost always be the default \code{.fcn = function(x) mean(x,na.rm=TRUE)}.}

\item{.i}{Character or character vector with the variable names that identify the individual cases. Note that setting any one of \code{.i}, \code{.t}, or \code{.d} will override all three already applied to the data, and will return data that is \code{pdeclare()}d with all three, unless \code{.setpanel=FALSE}.}

\item{.t}{Character variable with the single variable name indicating the time. \code{pmdplyr} accepts two kinds of time variables: numeric variables where a fixed distance \code{.d} will take you from one observation to the next, or, if \code{.d=0}, any standard variable type with an order. Consider using the \code{time_variable()} function to create the necessary variable if your data uses a \code{Date} variable for time.}

\item{.d}{Number indicating the gap in \code{.t} between one period and the next. For example, if \code{.t} indicates a single day but data is collected once a week, you might set \code{.d=7}. To ignore gap length and assume that "one period ago" is always the most recent prior observation in the data, set \code{.d=0}. By default, \code{.d=1}.}

\item{.uniqcheck}{Logical parameter. Set to TRUE to always check whether \code{.i} and \code{.t} uniquely identify observations in the data. By default this is set to FALSE and the check is only performed once per session, and only if at least one of \code{.i}, \code{.t}, or \code{.d} is set.}
}
\description{
WARNING: FOR NOW, ALWAYS SPECIFY THE PANEL STRUCTURE EVERY TIME YOU CALL THIS FUNCTION. PDECLARE STATUS IS DROPPED BY MOST FUNCTIONS.
}
\details{
Thesese functions perform the standard between and within transformations on panel data.
}
\examples{

data(SPrail)
#Calculate within- and between-route variation in price and add it to the data
SPrail <- SPrail \%>\%
    dplyr::mutate(within_route = within_i(price,.i=c('origin','destination')),
    between_route = between_i(price,.i=c('origin','destination')))

}
