% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_pd_methods.R
\name{pdeclare_methods}
\alias{pdeclare_methods}
\alias{mutate.tbl_pd}
\alias{select.tbl_pd}
\title{pdeclare methods}
\usage{
\method{mutate}{tbl_pd}(.data, ...)

\method{select}{tbl_pd}(.data, ...)
}
\description{
These are variants of existing functions that are designed to retain the \code{pdeclare} status of the object, as well as its \code{.i}, \code{.t}, and \code{.d} attributes.
}
\details{
If a function is not on this list, then you may need to re-\code{as_pdeclare} your object after using the function.
}
