% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
<<<<<<< HEAD
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
=======
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
>>>>>>> 7e2913474e048ac14379b7bdfedfc53f9c673542
}
\keyword{internal}
