% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{SPrail}
\alias{SPrail}
\title{2,000 Spanish train trips}
\format{A data frame with 2,000 rows and 9 variables:
\describe{
  \item{insert_date}{Date and time when ticket was paid for.}
  \item{origin}{Origin City}
  \item{destination}{Destination City}
  \item{start_date}{Date and time for train departure.}
  \item{end_date}{Date and time for train arrival.}
  \item{train_type}{Train service name.}
  \item{price}{Price of ticket in Euros.}
  \item{train_class}{Class of ticket: tourist, business, etc.. Variable in Spanish.}
  \item{fare}{Type of ticket fare.}
}}
\source{
\url{https://www.kaggle.com/thegurusteam/spanish-high-speed-rail-system-ticket-pricing}
}
\usage{
SPrail
}
\description{
This data set is a random subsample of a much larger database of trips taken on the Spanish High Speed Train Service (Renfe AVE).
}
\details{
All dates and times are European Central Time.

The larger data set from which \code{SPrail} was sampled was compiled and released under GPL-2 public license by Pedro Muñoz and David Cañones.
}
\keyword{datasets}
