% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_convert.R
\name{panel_convert}
\alias{panel_convert}
\title{Convert between panel data types}
\usage{
panel_convert(data, to, ...)
}
\arguments{
\item{data}{Data frame - a \code{pibble}, \code{tsibble}, \code{pdata.frame}, or \code{panel_data} object.}

\item{to}{Character variable set to \code{"pmdplyr", "pibble", "tbl_pb", "tsibble", "tbl_ts", "plm", "pdata.frame", "panelr"} or \code{"panel_data"} indicating the type/package to be converted to.}

\item{...}{Additional arguments to be sent to, respectively, \code{as_pibble()}, \code{tsibble::as_tsibble()}, \code{plm::pdata.frame()}, or \code{panelr::panel_data()}.}
}
\description{
This function takes panel data objects declared using \code{pmdplyr} (\code{pibble}/\code{tbl_pb}), \code{tsibble} (\code{tsibble}/\code{tbl_ts}), \code{plm} (\code{pdata.frame}), and \code{panelr} (\code{panel_data}) and converts to one of the other three formats for use with functions in those packages.
}
\details{
Any grouping will be lost. You must have the relevant package installed to convert to the type for that package. Conversions from \code{pdata.frame} will be improved if \code{sjlabelled} is also installed.

When using \code{panel_convert}, be aware of the requirements that each type has:

\tabular{lcccc}{
   Feature/Requirement \tab  \code{pibble}   \tab  \code{tsibble} \tab  \code{pdata.frame} \tab  \code{panel_data} \cr
                         \tab           \tab                \tab                 \tab \cr
   ID                    \tab \code{.i} \tab \code{key}     \tab \code{index[1]} \tab \code{id} \cr
   Time                  \tab \code{.t} \tab \code{index}   \tab \code{index[2]} \tab \code{wave} \cr
   Gap control           \tab \code{.d} \tab \code{regular} \tab No              \tab No  \cr
   ID must exist         \tab No        \tab No             \tab Yes             \tab Yes \cr
   Time must exist       \tab No        \tab Yes            \tab Yes             \tab Yes[1] \cr
   Only one ID variable[2]\tab No        \tab No             \tab Yes             \tab Yes \cr
   Unique identification \tab No        \tab Yes            \tab No[3]           \tab No[3]
 }

[1] \code{pdata.frame} does not require that time be provided, but if not provided will create it based on original ordering of the data. The \code{pdata.frame} option to set \code{index} equal to an integer for a balanced panel and have it figure out the rest by itself is not supported.

[2] Use \code{pmdplyr::id_variable()} to generate a single ID variable from multiple if one is required.

[3] \code{pdata.frame} and \code{panel_data} do not require that ID and time uniquely identify the observations on declaring the data, but functions in these packages may not work correctly without unique identification.

In addition to the above, be aware that the different packages have different requirements on which variable classes can be Time variables. \code{pmdplyr::time_variable()} can build an integer variable that will work in all packages.

You may run into some trouble if your data contains variables by the names \code{panel_convert_id}, \code{panel_convert_time}, \code{pibble_d}, or \code{panel_convert_regular}.
}
\examples{
# Only run examples if all the relevant packages are installed
if (sum(c("tsibble", "plm", "panelr") \%in\% utils::installed.packages()) == 3) {
  data(Scorecard)

  S_pibble <- as_pibble(Scorecard, .i = unitid, .t = year)
  S_tsibble <- tsibble::as_tsibble(Scorecard, key = unitid, index = year)
  S_pdata.frame <- plm::pdata.frame(Scorecard, index = c("unitid", "year"))
  S_panel_data <- panelr::panel_data(Scorecard, id = unitid, wave = year)

  # Make everything a pibble
  # Note variable order may be different coming from panelr, which changes variable order
  S_pibble
  panel_convert(S_tsibble, to = "pibble")
  panel_convert(S_pdata.frame, to = "pibble")
  panel_convert(S_panel_data, to = "pibble")

  # Make everything a tsibble
  panel_convert(S_pibble, to = "tsibble")
  S_tsibble
  panel_convert(S_pdata.frame, to = "tsibble")
  panel_convert(S_panel_data, to = "tsibble")

  # Now for pdata.frame
  head(panel_convert(S_pibble, to = "plm"))
  head(panel_convert(S_tsibble, to = "plm"))
  head(S_pdata.frame)
  head(panel_convert(S_panel_data, to = "plm"))

  # And finally panel_data
  panel_convert(S_pibble, to = "panelr")
  panel_convert(S_tsibble, to = "panelr")
  panel_convert(S_pdata.frame, to = "panelr")
  S_panel_data
}
}
